+++ src/router/type.ts
@@ -1,0 +1,7 @@
+import type { ComponentType } from 'react';
+
+export interface RouteConfig {
+  path: string;
+  Component: ComponentType;
+  text: string;
+}
